+++ lib/src/add_missing.dart
@@ -3,8 +3,8 @@ library add_missing;
 
 /// Add the missing columns for each of the rows of input [xs].  Fill with
 /// the [fill] value.
-List<Map<String,dynamic>> addMissing(List<Map<String,dynamic>> xs, {fill: null}) {
-  // get the missing columns for each row
+List<Map<String,dynamic>> addMissing(List<Map<String,dynamic>> xs, {dynamic fill}) {
+  // get all the unique columns from all rows
   var _columnNames = <String>{};
   for (var row in xs) {
     _columnNames.addAll(row.keys);